requirements.txt
tensorflow>=2.12.0
keras>=2.12.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.28.0
anthropic>=0.30.0
kaggle>=1.5.16
